Fixed Navigation and Settings
Fixed functionality with navigation so that when you leave a screen it dismisses instead of creating new instances of screens. Fixed settings so that a user can edit their social medias saved on the parse server, and that if they already have socials saved they appear in the settings screens.